+++ src/main/java/org/jabref/gui/entryeditor/SourceTab.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import java.util.Objects;
 
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.scene.Node;
 import javafx.scene.control.Tooltip;
 
@@ -46,6 +47,7 @@ public class SourceTab extends EntryEditorTab {
     private final BasePanel panel;
     private CodeArea codeArea;
     private BooleanProperty movingToDifferentEntry;
+    private BooleanProperty fieldsChanged = new SimpleBooleanProperty();
 
     public SourceTab(BasePanel panel, BibEntry entry, BooleanProperty movingToDifferentEntry) {
         this.mode = panel.getBibDatabaseContext().getMode();
@@ -65,6 +67,10 @@ public class SourceTab extends EntryEditorTab {
         new BibEntryWriter(formatter, false).writeWithoutPrependedNewlines(entry, stringWriter, type);
 
         return stringWriter.getBuffer().toString();
+    }
+
+    public BooleanProperty getFieldsChangedProperty() {
+        return fieldsChanged;
     }
 
     @Subscribe
@@ -166,6 +172,8 @@ public class SourceTab extends EntryEditorTab {
                 entry.clearCiteKey();
             }
 
+            boolean fieldsHaveChanged = !entry.getFieldNames().containsAll(newEntry.getFieldNames()) || !newEntry.getFieldNames().containsAll(entry.getFieldNames());
+
             // First, remove fields that the user has removed.
             for (Map.Entry<String, String> field : entry.getFieldMap().entrySet()) {
                 String fieldName = field.getKey();
@@ -200,6 +208,11 @@ public class SourceTab extends EntryEditorTab {
             }
             compound.end();
 
+            // notify about field changes
+            if (fieldsHaveChanged) {
+                fieldsChanged.setValue(true);
+            }
+
         } catch (InvalidFieldValueException | IOException ex) {
             // The source couldn't be parsed, so the user is given an
             // error message, and the choice to keep or revert the contents
@@ -223,5 +236,6 @@ public class SourceTab extends EntryEditorTab {
                 }
             }
         }
+        fieldsChanged.set(false);
     }
 }